feat(auth): add user registration, login, and JWT token handling

+++ app/app.py
@@ -1,24 +1,33 @@
-from fastapi import FastAPI,Form,File,UploadFile,Depends,HTTPException
-from app.db import Post,create_db_and_tables,get_async_session
-from contextlib import asynccontextmanager
+from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
+from app.schema import PostCreate, PostResponse, UserRead, UserCreate, UserUpdate
+from app.db import Post, create_db_and_tables, get_async_session, User
 from sqlalchemy.ext.asyncio import AsyncSession
-from dotenv import load_dotenv
-import os
+from contextlib import asynccontextmanager
 from sqlalchemy import select
 from app.images import imagekit
 import shutil
-import tempfile
+import os
 import uuid
+import tempfile
+from app.users import auth_backend, current_active_user, fastapi_users
 
-load_dotenv()
 
 @asynccontextmanager
-async def lifespan(app : FastAPI):
+async def lifespan(app: FastAPI):
     await create_db_and_tables()
     yield
 
 app = FastAPI(lifespan=lifespan)
 
+app.include_router(fastapi_users.get_auth_router(auth_backend), prefix='/auth/jwt', tags=["auth"])
+app.include_router(fastapi_users.get_register_router(UserRead, UserCreate), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_reset_password_router(), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_verify_router(UserRead), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_users_router(UserRead, UserUpdate), prefix="/users", tags=["users"])
+
+@app.post('/')
+async def home():
+    return {'message': 'Social Media App API Using FastAPI !'}
 
 @app.post('/upload')
 async def upload_file(
@@ -64,45 +73,54 @@ async def upload_file(
             os.unlink(temp_file_path)
         await file.close()
 
-
-@app.get('/feed')
+@app.get("/feed")
 async def get_feed(
-    session: AsyncSession = Depends(get_async_session)
+        session: AsyncSession = Depends(get_async_session),
+        user: User = Depends(current_active_user),
 ):
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts = [row[0] for row in result.all()]
+
+    result = await session.execute(select(User))
+    users = [row[0] for row in result.all()]
+    user_dict = {u.id: u.email for u in users}
 
     posts_data = []
     for post in posts:
         posts_data.append(
             {
                 "id": str(post.id),
+                "user_id": str(post.user_id),
                 "caption": post.caption,
                 "url": post.url,
                 "file_type": post.file_type,
                 "file_name": post.file_name,
-                "created_at": post.created_at.isoformat() 
+                "created_at": post.created_at.isoformat(),
+                "is_owner": post.user_id == user.id,
+                "email": user_dict.get(post.user_id, "Unknown")
             }
         )
 
-    return { "posts" : posts_data}
+    return {"posts": posts_data}
 
 
-@app.delete('/post/{post_id}')
-async def delete_post(post_id : str, session : AsyncSession = Depends(get_async_session)):
+@app.delete("/posts/{post_id}")
+async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_session), user: User = Depends(current_active_user),):
     try:
         post_uuid = uuid.UUID(post_id)
 
         result = await session.execute(select(Post).where(Post.id == post_uuid))
         post = result.scalars().first()
 
-        if not post :
-            raise  HTTPException(status_code=404, detail="Post Not Found !")
-        
+        if not post:
+            raise HTTPException(status_code=404, detail="Post not found")
+
+        if post.user_id != user.id:
+            raise HTTPException(status_code=403, detail="You don't have permission to delete this post")
+
         await session.delete(post)
         await session.commit()
 
-        return {"success": True, "message": "Post Deleted Successfully !"}
-
+        return {"success": True, "message": "Post deleted successfully"}
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
